requirements.txt
anybadge==1.7.0
astroid==2.5.6; python_version ~= '3.6'
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
cattrs==1.5.0; python_version >= '3.7'
cdktf-cdktf-provider-azurerm==0.1.85
cdktf-cdktf-provider-google==0.2.3
cdktf-cdktf-provider-kubernetes==0.2.62
cdktf==0.4.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==8.0.1
constructs==3.3.75; python_version >= '3.6'
decorator==5.0.9; python_version >= '3.5'
docker==5.0.0
gitdb==4.0.7; python_version >= '3.4'
gitpython==3.1.17
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==4.5.0; python_version >= '3.6'
isort==5.8.0; python_version >= '3.6' and python_version < '4.0'
jinja2==3.0.1; python_version >= '3.6'
jsii==1.30.0; python_version ~= '3.6'
lazy-object-proxy==1.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
markupsafe==2.0.1; python_version >= '3.6'
marshmallow-dataclass==8.4.1
marshmallow-enum==1.5.1
marshmallow==3.12.1; python_version >= '3.5'
mccabe==0.6.1
mypy-extensions==0.4.3
publication==0.0.3
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pylint-enums==0.0.5
pylint-gitlab==0.3.0
pylint==2.8.3; python_version ~= '3.6'
pynacl==1.4.0
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-slugify==5.0.2
pyyaml==5.4.1
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==4.0.0; python_version >= '3.5'
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4.0'
validators==0.18.2
websocket-client==1.0.1; python_version >= '3.6'
wrapt==1.12.1
yaspin==2.0.0
zipp==3.4.1; python_version >= '3.6'
